@@ -4,7 +4,8 @@
     "license": "MIT",
     "keywords": [
         "php",
-        "hyperf"
+        "hyperf",
+        "validate"
     ],
     "description": "Hyperf Validate",
     "autoload": {
@@ -18,12 +19,12 @@
         }
     },
     "require": {
-        "php": ">=7.3",
-        "hyperf/config": "^2.2",
-        "hyperf/db-connection": "^2.2",
-        "hyperf/di": "^2.2",
-        "hyperf/framework": "2.2.*",
-        "hyperf/validation": "^2.2"
+        "php": ">=8.0",
+        "hyperf/config": "~3.0",
+        "hyperf/db-connection": "~3.0",
+        "hyperf/di": "~3.0",
+        "hyperf/framework": "~3.0",
+        "hyperf/validation": "~3.0"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.0",
