@@ -33,7 +33,7 @@ class RequestHandle implements \DeathSatan\Hyperf\Validate\Contract\CustomHandle
     /**
      * {@inheritDoc}
      */
-    public function provide(object $current, AbstractValidate $validate, string $scene = null): array
+    public function provide(object $current, AbstractValidate $validate, array|null|string $scene = null): array
     {
         return $this->request->all();
     }
